requirements.txt
nba_api==1.10.2
pandas==2.3.3
psycopg2-binary==2.9.11
boto3==1.42.46
